Update example app to support attributes changes from Product Update view
